move protoc helpers into separate package so they don't end up linked into non-test code
